Finish refactor of 'Pseudonymizer' builder. Make all builders use refactored versions
