+++ .github/workflows/release.yml
@@ -49,7 +49,7 @@ jobs:
             bash -o pipefail -c "poetry version | cut -f 2 -d' '"
 
       - name: Bump version for developmental release
-        if: "! steps.check-version.outputs.tag"
+        if: '! steps.check-version.outputs.tag'
         run: |
           poetry version patch &&
           version=$(poetry version | awk '{ print $2 }') &&
