+++ src/dapla_pseudo/__init__.py
@@ -20,10 +20,9 @@ from importlib.metadata import version
 # Ref: https://github.com/python-poetry/poetry/issues/144#issuecomment-1488038660
 __version__ = version("dapla_toolbelt_pseudo")
 
-from dapla_pseudo.v1.api_models import Field
-from dapla_pseudo.v1.api_models import PseudoKeyset
 from dapla_pseudo.v1.client import PseudoClient
 from dapla_pseudo.v1.depseudo import Depseudonymize
+from dapla_pseudo.v1.models.core import PseudoKeyset
 from dapla_pseudo.v1.pseudo import Pseudonymize
 from dapla_pseudo.v1.repseudo import Repseudonymize
 from dapla_pseudo.v1.validation import Validator
@@ -34,6 +33,5 @@ __all__ = [
     "Depseudonymize",
     "Repseudonymize",
     "Validator",
-    "Field",
     "PseudoKeyset",
 ]
