+++ src/dapla_pseudo/types.py
@@ -8,9 +8,6 @@ import fsspec
 import gcsfs
 import pandas as pd
 
-from dapla_pseudo.v1.api_models import Field
-
-FieldDecl = str | dict[str, str] | Field
 BinaryFileDecl = t.Union[
     io.BufferedReader,
     fsspec.spec.AbstractBufferedFile,
