+++ src/dapla_pseudo/v1/__init__.py
@@ -1,9 +1,8 @@
 """Pseudo Service version 1 implementation."""
 
-from dapla_pseudo.v1.api_models import Field
-from dapla_pseudo.v1.api_models import PseudoKeyset
 from dapla_pseudo.v1.client import PseudoClient
 from dapla_pseudo.v1.depseudo import Depseudonymize
+from dapla_pseudo.v1.models.core import PseudoKeyset
 from dapla_pseudo.v1.pseudo import Pseudonymize
 from dapla_pseudo.v1.repseudo import Repseudonymize
 from dapla_pseudo.v1.validation import Validator
@@ -14,6 +13,5 @@ __all__ = [
     "Depseudonymize",
     "Repseudonymize",
     "Validator",
-    "Field",
     "PseudoKeyset",
 ]
